Rename autograd module to tensor
